Add files via upload
This is a simple rock, paper, scissors game with no GUI. The player just types their choice in the terminal and the computer plays it's move and spits out a smark alec comment depending on the results of the game. You can also play using Chinese in the SJB.py.